Correctly support "@parameters".
