--- src/main/java/com/sap/dsc/aas/lib/config/pojo/ConfigPlaceholder.java
@@ -1,38 +1,0 @@
-/* 
-  SPDX-FileCopyrightText: (C)2021 SAP SE or an affiliate company and aas-transformation-library contributors. All rights reserved. 
-
-  SPDX-License-Identifier: Apache-2.0 
- */
-package com.sap.dsc.aas.lib.config.pojo;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class ConfigPlaceholder {
-
-    private String name;
-    private String description;
-
-    @JsonCreator
-    public ConfigPlaceholder(@JsonProperty(required = true, value = "name") String name,
-        @JsonProperty(required = true, value = "description") String description) {
-        this.name = name;
-        this.description = description;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-}
