+++ src/main/java/com/sap/dsc/aas/lib/mapping/model/Parameter.java
@@ -1,0 +1,25 @@
+/* 
+  SPDX-FileCopyrightText: (C)2021 SAP SE or an affiliate company and aas-transformation-library contributors. All rights reserved. 
+
+  SPDX-License-Identifier: Apache-2.0 
+ */
+package com.sap.dsc.aas.lib.mapping.model;
+
+public class Parameter {
+
+    private final String name;
+    private final String description;
+
+    public Parameter(String name, String description) {
+        this.name = name;
+        this.description = description;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
