+++ src/test/java/com/sap/dsc/aas/lib/config/pojo/ParameterTest.java
@@ -16,8 +16,9 @@ import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.exc.MismatchedInputException;
 import com.sap.dsc.aas.lib.TestUtils;
+import com.sap.dsc.aas.lib.mapping.model.Parameter;
 
-public class ConfigPlaceholderTest {
+public class ParameterTest {
 
     private ObjectMapper objectMapper;
 
@@ -31,11 +32,11 @@ public class ConfigPlaceholderTest {
     void fromJsonString() throws JsonMappingException, JsonProcessingException {
         String input = "{\"name\": \"placeholderName\", \"description\": \"ui text\"}";
 
-        ConfigPlaceholder configPlaceholder = objectMapper.readValue(input, ConfigPlaceholder.class);
-        assertEquals("placeholderName", configPlaceholder.getName());
-        assertEquals("ui text", configPlaceholder.getDescription());
+        Parameter parameter = objectMapper.readValue(input, Parameter.class);
+        assertEquals("placeholderName", parameter.getName());
+        assertEquals("ui text", parameter.getDescription());
 
-        assertThrows(MismatchedInputException.class, () -> objectMapper.readValue("{}", ConfigPlaceholder.class));
+        assertThrows(MismatchedInputException.class, () -> objectMapper.readValue("{}", Parameter.class));
     }
 
 }
